requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
imbalanced-learn>=0.10.1
joblib>=1.2.0
kiwisolver==1.4.8
matplotlib>=3.7.1
numpy>=1.24.3
nvidia-nccl-cu12==2.25.1
packaging==24.2
pandas>=2.0.2
pillow==11.1.0
psutil==6.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn>=1.2.2
scipy==1.15.1
seaborn>=0.12.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.5.0
tzdata==2025.1
xgboost>=1.7.5
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0
optuna>=3.5.0
discord-webhook>=1.0.0
